deployment_engine/requirements.txt
fastapi
uvicorn
pydantic
python-jose[cryptography]
python-multipart
requests

# Azure dependencies
azure-identity
azure-mgmt-resource
azure-mgmt-compute
azure-mgmt-storage
azure-mgmt-network

# AWS dependencies
boto3

# GCP dependencies
google-cloud-deployment-manager
google-cloud-storage
google-auth

# For JWT token validation
pyjwt

